SupervisedLearning/requirements.txt
###### Requirements without Version Specifiers ######
python-javabridge>=4.0.0
python-weka-wrapper3

